Parse MediaWiki dump XML and store page names in the index
This task is mostly bound by the network speed, so the fast that quick-xml
might be slower of faster than awk is irrelevant. Further benchmarks might
be needed later on if the mediawiki dumps are loaded locally, such as on
Toolforge.

Note: page names alone for frwiki account for 326 Mo uncompressed, or 102 Mo
compressed with gzip --best. It's probably a good idea to compress using zstd
at build time and to decompress in RAM at query time (< 2 s with gzip).