Implement /query very naively
This opens the index multiple times and works on it for every single query,
plus the index comes in the form of a filename instead of a BufRead, which is
not desirable either.

Performance is actually suprisingly good when I/Os are fast (eg. SSD), but it
is abysmal when their aren't (eg. NFS under load).

This is quite fine for personal use, but doesn't fit the bill for toolforge.